OpenGeoSys BHE Simulation Tutorial
Borehole heat exchanger plays an important role in geothermal energy exploration. Besides the rule of thumb, there are more and more technicians tend to use simulation tools in the project evaluation and design. This manual is presented to give a tutorial in simulating the performance of BHE through OGS-6. By using this tutorial, anyone could implement the simulation of BHEs considering variation of geothermal properties, borehole characteristics and even layout of multi-boreholes. This tutorial will help users in coping with the problem of BHE in application and operation.